+++ src/Exceptions/ModelConfigurationException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Codeart\OpensearchLaravel\Exceptions;
-
-class ModelConfigurationException extends \Exception implements OpenSearchException
-{
-
-}