+++ src/Exceptions/ModelException.php
@@ -2,11 +2,9 @@
 
 namespace Codeart\OpensearchLaravel\Exceptions;
 
-use JetBrains\PhpStorm\Pure;
-
 class ModelException extends \Exception implements OpenSearchException
 {
-    #[Pure] public function __construct($message, $code = 500)
+    public function __construct($message, $code = 500)
     {
         $previous = null;